move nativeclasses out of init macro
