+++ Cargo.toml
@@ -14,3 +14,4 @@ trace_chrome = ["trace", "bevy/trace_chrome"]
 gdnative = "0.10"
 bevy = {version = "0.8", default-features = false, features = ["bevy_asset"]}
 anyhow = "1.0.58"
+lazy_static = "1.4.0"
